% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_helpers}
\alias{diff_helpers}
\alias{diff_css}
\alias{diff_to_html}
\alias{view_diff}
\title{Formatting and viewing diffs}
\usage{
diff_css(palette = "vscode")

diff_to_html(x)

view_diff(x, palette = "vscode")
}
\arguments{
\item{palette}{palette name passed to \code{\link[cli:ansi_html_style]{cli::ansi_html_style()}}}

\item{x}{diff, as returned by \code{\link[=diff_text]{diff_text()}}}
}
\value{
\code{\link[=diff_css]{diff_css()}} returns CSS to style diff output, \code{\link[=diff_to_html]{diff_to_html()}}
returns HTML as returned by \code{\link[htmltools:HTML]{htmltools::HTML()}}, and \code{\link[=view_diff]{view_diff()}} renders
the diff in the RStudio Viewer (interactive sessions only).
}
\description{
These functions prepare diffs for inclusion in HTML documents or viewing
in interactive sessions
}
\examples{
diff_text(letters, sub("d", "D", letters)) |>
  diff_to_html() |>
  cat()

}
