% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_text}
\alias{diff_text}
\title{Compare two character vectors}
\usage{
diff_text(a, b)
}
\arguments{
\item{a, b}{character vectors to compare}
}
\value{
Returns a \href{https://www.gnu.org/software/diffutils/manual/html_node/Example-Unified.html}{unified diff}
as a character vector.
}
\description{
This function uses \href{https://git-scm.com/docs/git-diff}{\verb{git diff}} to
compare two character vectors.
}
\examples{
diff_text(letters, sub("d", "D", letters)) |>
  paste(collapse = "\n") |>
  write("")
}
