% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{estimate_line_length}
\alias{estimate_line_length}
\title{Estimate text width (measure)}
\usage{
estimate_line_length(font, fontsize = c("10pt", "11pt", "12pt"), wrap = 66)
}
\arguments{
\item{font}{name of font, passed to \verb{\\setmainfont\{\}}}

\item{fontsize}{fontsize, passed to \verb{\\documentclass} options}

\item{wrap}{desired number of characters per line, passed to \code{\link[=sample_text]{sample_text()}}.}
}
\value{
Returns a named list with average ("mean") and standard deviation
("sd") of desired text width
}
\description{
This function typesets bits of text to estimate the width of a textblock
that averages \code{wrap} characters wide. This is useful for back-calculating
page dimensions to suit the optimal line length of 66 characters (or whatever
your project calls for).
}
\details{
This rather brutish function uses \href{https://www.latex-project.org}{LuaLaTeX}
with \href{https://ctan.org/pkg/fontspec}{fontspec} and
\href{https://docs.ropensci.org/pdftools/}{pdftools} for R, and it assumes that
your TeX installation is configured to access fonts by name. It makes no
attempt to check your system for required resources, handle errors, or
protect against injection of malicious code.
}
