% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{fit_layouts_to_text_width}
\alias{fit_layouts_to_text_width}
\title{Scale generic layouts}
\usage{
fit_layouts_to_text_width(
  target_text_width,
  fontsize = c("10pt", "11pt", "12pt"),
  max_page_height = 8.5,
  max_page_width = 5.5
)
}
\arguments{
\item{target_text_width}{desired textblock width in inches}

\item{fontsize}{base font size for document, used to estimate the number of
lines of text the textblock can accommodate}

\item{max_page_height, max_page_width}{maximum page dimensions (default
assumes layout is printed on letter paper folded in half)}
}
\value{
Returns \link{layouts} plus three additional columns: \code{font_size},
\code{min_lines}, & \code{max_lines}.
}
\description{
Resize generic layouts to fit a desired textblock width, and filter out
layouts that are too large for the specified stock paper.
}
\examples{
str(fit_layouts_to_text_width(4))
}
