% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{name_ratios}
\alias{name_ratios}
\title{Match numeric proportions to named harmonious ratios}
\usage{
name_ratios(h, w = 1)
}
\arguments{
\item{h, w}{pairwise height & width vectors representing input proportions}
}
\value{
Returns a character vector of ratio names or NA for input
proportions that lack a match in \code{letterblasteR::ratios}
}
\description{
This function simply looks up the name of the harmonious ratio that matches
the input proportion.
}
\examples{
name_ratios(3, 2)
name_ratios(3.0001, 2)
}
\seealso{
\code{vignette("ratios")}
}
