% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_download}
\alias{pg_download}
\alias{pg_download_text}
\title{Download Project Gutenberg content}
\usage{
pg_download_text(url, zip = FALSE, encoding = "UTF-8")
}
\arguments{
\item{url}{URL for Project Gutenberg file}

\item{zip}{logical indicating if file is zipped}

\item{encoding}{assumed encoding of source file}
}
\value{
Returns content as a UTF-encoded character vector.
}
\description{
These functions download content from Project Gutenberg, tidying intermediate
files along the way.
}
