% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_metadata}
\alias{pg_metadata}
\title{Download Project Gutenberg metadata for a single work}
\usage{
pg_metadata(pgid)
}
\arguments{
\item{pgid}{Project Gutenberg ID (i.e.
\verb{https://www.gutenberg.org/ebooks/<PGID>}), a single integer value}
}
\value{
Returns a \link[tibble:tibble-package]{tibble} of RDF triples where
\emph{subject}, \emph{predicate}, and \emph{object} have been renamed to \emph{key}, \emph{name},
and \emph{value}, discarding the RDF structure in favor of a simpler "long"
format that can be aggregated and reshaped easily through tidyverse
conventions. Note that recursive joins may be required to capture nested
attributes.
}
\description{
\href{https://www.gutenberg.org}{Project Gutenberg} provides metadata in \href{https://en.wikipedia.org/wiki/Resource_Description_Framework}{RDF format} for
each of their works. This function downloads the file for the requested work
and transforms it into a tidy data frame.
}
\section{Not for scraping!}{
Do not use this function to scrape metadata in bulk. Project Gutenberg
provides \href{https://www.gutenberg.org/ebooks/offline_catalogs.html}{offline catalog downloads} for that
purpose. Be a courteous data scientist and take steps to avoid calling this
function repeatedly in your code -- particularly during development. Consider
using the \href{https://memoise.r-lib.org}{memoise} package in your code and/or
\href{https://quarto.org/docs/computations/caching.html}{caching} in your
Rmarkdown and Quarto documents.
}

\references{
See \href{https://docs.ropensci.org/rdflib/articles/rdf_intro.html}{A tidyverse lover's intro to RDF} for a great
introduction to the RDF format.
}
