% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{plot_layouts}
\alias{plot_layouts}
\title{Visualize harmonious layouts}
\usage{
plot_layouts(x)
}
\arguments{
\item{x}{subset of \code{layouts}}
}
\value{
Returns a ggplot.
}
\description{
This function plots harmonious layouts, using transparency to overlay
multiple similar layouts at a time. It is meant to give an overview of
layout options and to home in on a subset of candidates for a project.
}
\examples{
plot_layouts(layouts)

layouts |>
  subset(
    layouts$page_ratio \%in\% c("fourth", "fifth") &
      dplyr::between(ratios[layouts$text_ratio], 1.2, 1.6) &
      dplyr::between(layouts$text_pct, 0.5, 0.75)
  ) |>
  plot_layouts()
}
