% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\docType{data}
\name{ratios}
\alias{ratios}
\title{Harmonious ratios}
\format{
A named vector
}
\usage{
ratios
}
\description{
A simple named vector of "harmonious" ratios.
}
\details{
Ratios fall into the following categories:
\itemize{
\item \href{https://en.wikipedia.org/wiki/Chromatic_scale}{chromatic} ratios derived
from western music theory
\item \href{https://en.wikipedia.org/wiki/Golden_ratio}{golden ratio}
\item "triple" and "double octave" ratios -- so set includes simple ratios of
1:1 (unison), 1:2 (octave), 1:3, and 1:4
}

All ratios (except for 1:1) appear in pairs, e.g. "major third" (1.25) and
"inverse major third" (0.8). Ratios greater than one imply a \emph{portrait}
orientation, and ratios less than one imply a \emph{landscape} orientation. Vector
is sorted from square (unison) to the tallest/widest proportions.
}
\examples{
cbind(ratios)
}
\seealso{
\code{vignette("ratios")}
}
\keyword{datasets}
