% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{round_ratios}
\alias{round_ratios}
\title{Round proportions to nearest harmonious ratios}
\usage{
round_ratios(
  h,
  w = 1,
  target = c("nearest", "taller", "wider"),
  tolerance = 0.01
)
}
\arguments{
\item{h, w}{pairwise height & width vectors representing input proportions}

\item{target}{whether to identify nearest harmonious ratio (default) or to
limit to solutions where the input proportion may only be taller/wider
than the harmonious ratio, see Details & Examples}

\item{tolerance}{only consider harmonious ratios within ±tolerance of input
ratios, in degrees}
}
\value{
Returns a named vector of rounded proportions. Proportions that are
not within ±tolerance of a harmonious ratio with be \code{NA}.
}
\description{
Given height & width vectors representing input proportions, this function
identifies the nearest harmonious ratio.
}
\details{
Input proportions are compared to harmonious ratios after converting to
angles with \code{\link[=degrees]{degrees()}}. Rounding only considers harmonious ratios that are
within ±tolerance (in degrees) of the input proportion.

By default, the nearest harmonious ratio is returned, but the \code{target}
argument can be used to favor actual proportions that run tall or wide. When
making page layout decisions, it can be useful to prefer proportions along
the spine that are a little wider than the ideal ratio. Once bound, optical
foreshortening along the spine tends to bring the perceived ratio back into
harmony.
}
\examples{
round_ratios(3, 2)

round_ratios(2.9999, 2)
round_ratios(3.0001, 2)

round_ratios(c(2.9999, 3.0001), 2, tolerance = 1e-6)

round_ratios(c(2.9999, 3.0001), 2, target = "taller")
round_ratios(c(2.9999, 3.0001), 2, target = "wider")
}
\seealso{
\code{vignette("ratios")}
}
