% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{score_ratios}
\alias{score_ratios}
\title{Score how well a proportion matches a harmonious ratio}
\usage{
score_ratios(h, w = 1, ...)
}
\arguments{
\item{h, w}{pairwise height & width vectors representing input proportions}

\item{...}{optional arguments passed to \code{round_ratios()}}
}
\value{
Returns a named numeric vector of scores in [0, 1]
}
\description{
This function identifies the nearest harmonious ratio with a call to
\code{round_ratios()} and calculates the difference in area between the actual
proportion and the ideal. The returned score describes the percent overlap
of the two rectangles.
}
\examples{
score_ratios(3, 2)
score_ratios(3 + 1e-4, 2)
score_ratios(3 - 1e-4, 2)
score_ratios(4)
score_ratios(5)
}
