% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{to_tidy_md}
\alias{to_tidy_md}
\title{Convert document to Pandoc's Markdown}
\usage{
to_tidy_md(
  x,
  input_format,
  shift_heading_level_by = 0L,
  default_image_extension = "jpg",
  dpi = 300,
  ...
)
}
\arguments{
\item{x}{path to single file or document text as a character vector}

\item{input_format}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#general-options}{\code{--from}}}

\item{shift_heading_level_by}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#reader-options}{\code{--shift-heading-level-by}}}

\item{default_image_extension}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#reader-options}{\code{--default-image-extension}}}

\item{dpi}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#general-writer-options}{\code{--dpi}}}

\item{...}{additional Pandoc options added to the \code{args} parameter of
\code{\link[pandoc:pandoc_convert]{pandoc::pandoc_convert()}}}
}
\value{
Returns a character vector of tidy markdown
}
\description{
Convert document to Pandoc's Markdown
}
