% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterblasteR-package.R
\docType{package}
\name{letterblasteR-package}
\alias{letterblasteR}
\alias{letterblasteR-package}
\title{letterblasteR: Utilities for Proofreading and Typesetting Books}
\description{
The letterblasteR package provides tools for proofreading public domain texts from Project Gutenberg and typesetting print-ready books.
}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
