% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterblasteR-package.R
\docType{package}
\name{letterblasteR-package}
\alias{letterblasteR}
\alias{letterblasteR-package}
\title{letterblasteR: Utilities for Proofreading and Typesetting Books}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The letterblasteR package provides tools for proofreading public domain texts from Project Gutenberg and typesetting print-ready books.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LetterblasterPress/letterblasteR}
  \item \url{https://letterblasterpress.github.io/letterblasteR/}
  \item Report bugs at \url{https://github.com/LetterblasterPress/letterblasteR/issues}
}

}
\author{
\strong{Maintainer}: Justin Brantley \email{fascinatingfingers@icloud.com}

}
\keyword{internal}
