% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_download}
\alias{pg_download}
\title{Download Project Gutenberg content}
\usage{
pg_download(url, dest_dir, zip = FALSE)
}
\arguments{
\item{url}{URL for Project Gutenberg file}

\item{dest_dir}{path to save (uncompressed) files}

\item{zip}{logical indicating if source file is zipped}
}
\value{
Returns path to \code{dest_dir} invisibly.
}
\description{
Download (possibly zipped) content from Project Gutenberg to a local
directory.
}
