% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{sample_text}
\alias{sample_text}
\title{Generate sample text}
\usage{
sample_text(n, wrap = NULL)
}
\arguments{
\item{n}{number of sentences to return, or number of lines if \code{wrap} is not
\code{NULL}}

\item{wrap}{column to wrap text at, optional

If \code{wrap} is specified, text is wrapped exactly at that column, disregarding
language. Lines are rearranged at random after filtering out lines with
leading or trailing space.}
}
\description{
Generate a sample of "Harvard sentences", wrapping to an exact column width
if desired.
}
\examples{
sample_text(3)
sample_text(3, wrap = 42)
}
\seealso{
\link[stringr:stringr-data]{stringr::sentences}
}
