% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{to_tidy_md}
\alias{to_tidy_md}
\title{Convert document to Pandoc's Markdown}
\usage{
to_tidy_md(x, input_format, to = "markdown", ...)
}
\arguments{
\item{x}{path to single file or document text as a character vector}

\item{input_format}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#general-options}{\code{--from}}}

\item{to}{passed to Pandoc option
\href{https://pandoc.org/MANUAL.html#general-options}{\code{--to}}}

\item{...}{additional Pandoc options added to the \code{args} parameter of
\code{\link[pandoc:pandoc_convert]{pandoc::pandoc_convert()}}}
}
\value{
Returns a character vector of tidy markdown
}
\description{
Convert document to Pandoc's Markdown
}
\examples{
# TODO --shift-heading-level-by
# TODO --default-image-extension
# TODO --dpi
# TODO --extract-media

}
